Update patch set 1
Patch Set 1:

(3 comments)

Hey Connor, this is draft of a .hal file for the boot_control so I have something I can start playing with in update_engine. Take it, or drop it and re-write something else. Note: if you download this patch, make changes to it while keeping the Change-Id in the commit message, repo upload will upload it as another patch of this CL... it doesn't matter who the author was of the previous patch. Feel free to do so.

I don't think we need to enforce any call flow. I think every call flow is valid.

I didn't include the HAL in the Android.bp in the root of this git repo (I understand that that's how you declare your HAL public). To compile this you need to include this line in the hardware/interfaces/Android.bp:

    "boot_control/1.0",

enjoy!

Patch-set: 1
Reviewer: Gerrit User 1077068 <1077068@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0