wifi: Move WifiLegacyHal to a separate namespace
To prevent typename conflicts between the types in "hardware_legacy/wifi_hal.h"
and the HIDL interface. Move all the legacy HAL types under a separate
namespace (legacy_hal).
This is needed for Nan data types. Some of the typenames are exactly the
same in the legacy HAL header file and the HIDL interface.

Related changes,
a. Remove the use of |HalTool| class. This causes compilation failures
because it tries to include wifi_hal.h as well (but, that is now wrapped
inside a namespace).
b. Removed the usage of wifi_status_util in WifiLegacyHal. The
|legacyErrorToString| conversion is done in the HIDL object.

Bug: 32242225
Test: Compiles
Change-Id: I23dc704fbde5318b01fc1d5f41c811551debdd96