Update patch set 1
Patch Set 1:

Sepolicy allows access to the block device. What's blocking the access to the "misc" device (which is the standard Android-recommended way to implement A/B) and access to some other block devices in the case of  the Pixel phones HAL implementations is the fact that block device has permission:
  brw------- 1 root root

This is not a SELinux denial, is a plain linux user ACL denial.

Patch-set: 1