wifi: Implement chip mode combinations
This is pretty hardcoded for now. We expose 2 chip modes:
1. STA mode: 1 STA, 1 P2P and 1 NAN iface.
2. AP mode: 1 AP iface.

Implement the chip mode configuration related HIDL methods in
WifiChip and integrate WifiModeController.

TODO: Need to change the order of some operations in IWifi.start(). The
legacy HAL function to start the HAL needs to be called once the chip if
reconfigured. So, that needs to be moved from IWifi.start() to
IWifiChip.configureChip().

Bug: 31997422
Test: Compiles
Change-Id: I461687d7ee92398bc47321e1baca609db65c7991