Update patch set 6
Patch Set 6:

I'm not sure I understand the security model here. My first guess is that wifi_hal_legacy will start up as root, and transition to wifi if/when initializeAndDropPrivileges() is called (and succeeds).

But it's not obvious that this is guaranteed to happen. And I don't know what other methods might be called before privileges are dropped.

Is there a doc or bug that explains the security story?

Patch-set: 6
Label: Verified=0