Update patch set 2
Patch Set 2:

> > > Why exactly do you want this? The officially recommend logging
 > > API
 > > > for C++ is android-base/logging.h
 > >
 > > Well to clarify - it depends for what code; android/log.h is a C
 > > API in the NDK, log/log.h C API for VNDK, android-base/logging.h
 > > may or may not have a stable ABI. Since these are just tests that
 > > are built with the platform I guess it doesn't really matter, but
 > > you could argue to use android/log.h because it has a stable ABI.
 > > Wondering if that was the background here.
 > 
 > If it doesn't really matter, then I think this is the quickest fix
 > to allow the HIDLize IBase CLs to merge.
 > 
 > We can change to android-base/logging.h later. I'll file a bug to
 > track this.

It wasn't clear to me why you needed this change in the first place. If the IBase change is blocked on this, please file a bug explaining what we need and why, so we can pick the right logging interface to use for these interfaces. We should probably also think about some guidelines for interfaces in general and add it to our C++ guide.

Patch-set: 2