Update patch set 6
Patch Set 6:

Martijn is taking a look at b/33173166.

The bug b/33173166 is not *so* related to this CL though, because IBase's can still be passed around; it just needs to be re-wrapped because of the mysterious bug in b/33173166. IMO, b/33173166 is totally unrelated, but just happened to be discovered by this change.

Patch-set: 6