wifi: Add legacy HAL stubs
This is a port of all the existing stub functions from HalTool. Will
deprecate remove HalTool once framework switches over to HIDL.
These stubs are important because the vendor implemnetations don't fill
a function pointer in the table (when |init_wifi_vendor_hal_func_table| is
invoked) if they don't implement that feature.

Bug: 33540636
Test: Gtests
Change-Id: Ia41c4faa30b517731a162f3999d64a3f2c34e058