Update patch set 3
Patch Set 3:

> Realized that this radio work scheduling is iface specific. So,
 > moved these methods to IWifiStaIface since that is the one we'll be
 > interested in.
 
The rest of the change looks good, but I'm not sure I understand this part. I _do_ understand why PS5 moves the methods off of ISupplicant. But the choice of ISupplicantStaIface, rather than ISupplicantIface, is not obvious to me.

At least in the abstract, I don't see why the radio work methods should require a STA iface. It seems like these might be useful on a P2P iface, or perhaps even an AP iface.

Is the concern that the radio work methods would misbehave, if applied to P2P or AP ifaces? Or is this some other reason for placing the methods on StaIface, rather than (base) Iface?

Patch-set: 3