Update patch set 5
Patch Set 5:

Yep. There isn't anything in wpa_Supplicant which limits the functionality of this command for P2p  ifaces.
But, there are a couple of challenges with the HIDL implementation:
1. There isn't a base ISupplicantIfaceCallback object. So, we'll need to create one just for this method or duplicate these callbacks in both ISupplicantP2pIfaceCallback & ISupplicantStaIfaceCallback.
2. I'm not sure how this command needs to be handled for devices like angler where the P2P iface is a virtual iface on top of the STA iface (https://b.corp.google.com/issues/31755813#comment3).

Also, we don't really have any conflicts between android framework & wpa_supplicant for P2p operations. wificond is only interested in STA operations. wpa_supplicant is fully in control of P2P iface. So, there really isnt a need for these to be added for P2P ifaces.

Patch-set: 5
Label: Verified=0