Update patch set 1
Patch Set 1:

(1 comment)

Regarding threading documentation: I don't think we need to block on b/33552618. I think we can tackle the documentation that's relevant to this change here. Then, use the b/33552618 just for the parts that are mentioned in b/33552618.

In fact, I believe it's best to document the model now, while we're both thinking about it. (Threading is hard. Paging in threading details is harder.)

Patch-set: 1
Label: Verified=0