Cell location information in registration info
The data registration and voice registration states contain
large swaths of location information. While it seemed at first
blush we could remove that information, having synchronized
knowledge of when a registration has resulted in a cell change
provides valuable infomation. Cleanup registration response
using the same cell location structure that is being used
elsewhere in the interface to ensure that this method
always keeps sufficiently up-to-date information, without the
need to increasingly bloat it.

Test: Verified with updated qcril
Change-Id: Ib00063b6c3eed04da3f413c9dc1bcdff4750140e