Create HIDL HAL for WiFi Keystore
The WiFi keystore HAL should run in the system image, and will
provide an interface for wpa_supplicant (running on the vendor
partition) to call in order to perform keystore operations.
This service will run as the WiFi UID so that it can access all
credentials that WiFi processes would have.  As such, this
process can only be usable for WiFi and is not a generic HAL
for the keystore interface.

Bug: 34603782
Test: Compile only -- builds cleanly, but still WIP
Change-Id: Iac353c6dd182a09a11e87de5b0703e67f8e31ca8