Update patch set 2
Patch Set 2:

(please see b/36069735)
I understand that bluetooth team do the development in AOSP and they want to use the configstore (b/34351833). I have no idea about previous decision not to move to AOSP, but I think the confgstore will benefit others too if we do so.

I squashed commits for two reasons in addition to Jiyong’s suggestions; (1) reducing the overhead related to the code reviews and (2) there are some commits crossing various other components irrelevant to configstore making it impossible to cherry-pick as it is. I think (1) is not a problem for hardware/interfaces/configstore because you can give +2 for all at once. For (2), I suggest to cherry-pick ‘as it is’ for configstore-specific commits and ‘with squashing’ for other commits (minor modifications to Android.{bp|mk}). What about this plan ?

Patch-set: 2
Reviewer: Gerrit User 1143592 <1143592@85c56323-6fa9-3386-8a01-6480fb634889>