Update patch set 9
Patch Set 9:

I tried to pass with meaningful params down with 'set' cell broadcasting config apis. The result is: when sim card is not in, the 'set' with meaningful params is not success (it returns an error code INVALID_MODE_STATE or REQUEST_NOT_SUPPORTED, which may be a bug from QC so I marked it in the doc shared with QC); when sim card is in, the 'set' with meaningful params is success. 

I think putting 'get', 'set', 'get' into one unit test would be meaningful when 'set' is success, and 'set' with meaningful params can be success when sim card is in. Therefore, I have not implemented the unit test with combination of 'get', 'set' and 'get' at this stage that sim card is absent.

Patch-set: 9
Reviewer: Gerrit User 1131623 <1131623@85c56323-6fa9-3386-8a01-6480fb634889>