Update patch set 2
Patch Set 2:

I see. Yes, It's good to have it on AOSP. OK. I created b/38009470 for this. BTW yim@, zhuoyao@ and me talked about the possibility to merge the throughput test and latency test which depends the HAL. I'd like to have a short hangout meeting to sort it out. How about that?

Patch-set: 2