NFC: vts: Add commands to reset the NFC controller before other commands
Bandwidth and WriteInvalidAndThenValidCommand test assumes that the NFC
controller will be initialized before the test. This is not necessarily
true for all devices. Send CORE_RESET_CMD and CORE_INIT_CMD before
creating a loop-back connection.

CORE_RESET_RSP doesn't necessary return the configuration that was sent
in CORE_RESET_CMD. For instance, if CORE_RESET_CMD requested keeping
the configuration, and if the controller doesn't support that, the
configuration will be resetted and CORE_RESET_RSP will reflect that.

Bug: 38267593
Test: run vts
Change-Id: I65ae9746054ca767b928bdd461faa02f03719aea
Signed-off-by: Ruchi Kandoi <kandoiruchi@google.com>