NFC: vts: Bandwidth test has been modifed to reset the NFC controller
Bandwidth and WriteInvalidAndThenValidCommand test assumes that the NFC
controller will be initialized before the test. This is not necessarily true for
all devices. Send CORE_RESET_CMD and CORE_INIT_CMD before creating a loop-back
connection.

CORE_RESET_RSP doesn't necessary return the configuration that was sent
in the command. For instance, if CORE_RESET_CMD requested not resetting
the configuration, and the controller doesn't support that then the
response would reflect configuration reset status.

Bug: 38267593
Test: run vts
Change-Id: I65ae9746054ca767b928bdd461faa02f03719aea
Signed-off-by: Ruchi Kandoi <kandoiruchi@google.com>