Update patch set 7
Patch Set 7: Presubmit-Ready+1

The reason this makes me nervous is that there are other devices which are working with the code as it currently is.  In other words, if there is a partial write, WriteSafely will retry until it succeeds.  These platforms don't care if there is more than one write.

I am not sure that the current code can guarantee a single write call, and I think it will be hard to debug if partial packets are written.

Patch-set: 7
Reviewer: Gerrit User 1076123 <1076123@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1