DO NOT MERGE Docs: Fix HIDL doc comments for parse errors and warnings.
Conforms doc comment blocks to match HIDL style so they can
be parsed by hidl-doc.
To note, this doesn't mean all the HAL files are properly
formatted for docs---some files are using '//..' which hidl-doc
does not recognize as a doc block. This change just fixes doc
parse errors and warnings.

Bug: 34850358
Test: hidl-doc -i ./hardware/interfaces/
Change-Id: Icf81bafa670599d3413f347340b5b8cfdfa43944