Update patch set 1
Patch Set 1:

The hidl-doc tool just generates HTML docs from HAL files. It didn't create this changelist, I just wanted to fix some of the things it found.

Some things are warnings. If a method takes a argument and that does not have a corresponding doc param, hidl-doc will warn you can carry on. It will document the param as an empty string, but I'd argue that an empty doc string is not very helpful and thus warrants a warning.

Something like /* vs /** is much more ambiguous. As discussed in b/36453077, there is a place for comments and documentation in HAL files. The prime example is the license that appears at the top of every file---not a doc. And I found in numerous files, for example ISupplicantStaNetwork.hal, there's a comment block noting where the getters and setters start. These aren't docs. The '/**' makes it explicit that this is intended as documentation. The '/**' marker is the convention that javadoc and doxygen use.
hidl-doc handles /* blocks fine (it just strips them out). But in the case of ISupplicantStaNetwork.hal, there are two consecutive /** blocks. This is an error because the first block is not intended as documentation.

Patch-set: 1
Reviewer: Gerrit User 1125758 <1125758@85c56323-6fa9-3386-8a01-6480fb634889>