Update patch set 3
Patch Set 3:

> Xin, this's for vts change in colocated projects which don't have
 > oreo-vts-dev branch. to see this change in oreo-vts-dev, where this
 > needs to be merged to (oreo-release)?

No, changes should not land on -release on AOSP (most of these are considered as read-only, especially platform ones, e.g. <dessert>-release).

If the goal is to share this change with partners or third parties, the CL would be sufficient; if the goal is to have outside developers to be able to 'repo sync' and get the code, please hold for now and cherry-pick to oreo-vts-dev after b/64938555.

Patch-set: 3
Reviewer: Gerrit User 1086523 <1086523@85c56323-6fa9-3386-8a01-6480fb634889>