Update patch set 1
Patch Set 1:

Chih-hung, can you comment on this? Will an unknown email address cause any issues? I think that this would not suggest it as a reviewer, which is part of the problem with groups (in that they are unknown to gerrit as an account, and thus not selectable during find-owners).

Patch-set: 1
Reviewer: Gerrit User 1010056 <1010056@85c56323-6fa9-3386-8a01-6480fb634889>