Update patch set 4
Patch Set 4: Presubmit-Verified+1

=== Finished presubmit run: L77500000101831278 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/479944?revision=4




TreeHugger finished with: 11 passed.





WARNING: The tip-of-tree(4331725) for git_master/errorprone_fastbuild3e_linux may be broken.
The presubmit build passes with the last-known-good-build(4331579), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



WARNING: The tip-of-tree(4331725) for git_master/gce_x86_phone-userdebug_fastbuild3c_linux may be broken.
The presubmit build passes with the last-known-good-build(4331579), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.







WARNING: The tip-of-tree(4331725) for git_master/marlin-userdebug_fastbuild_linux may be broken.
The presubmit build passes with the last-known-good-build(4331491), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




WARNING: The tip-of-tree(4331725) for git_master/sailfish-userdebug_fastbuild3_linux may be broken.
The presubmit build passes with the last-known-good-build(4331579), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



I'm your biggest fan!

Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 4
Tag: autogenerated:TreeHugger:Verify-Pass
Reviewer: Gerrit User 1062513 <1062513@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Verified=+1