Move libhwminijail into configstore directory.
This library is only used by configstore, and since it doesn't
match a root package of a hidl interface, it messes up the
update-makefiles script. Perhaps we can move it into libhidl
or another library if more HALs want to use it--it's definitely
not hardware/interfaces or configstore specific in its
implementation.

It should also probably be consolidated with libavservices_minijail.

Test: none
Change-Id: I7c17292183d5cc1ff534ea51ac7f161081ee4f5c