Update patch set 1
Patch Set 1:

> (1 comment)

Sorry for missing this in my review. Seems like increasing the buffer size is indeed the simplest solution.

A possible alternative is to spawn a thread that keeps reading from the pipe until the call to 'debugDump' returns. This way we ensure that the HAL has finished writing, and we can safely close the pipe (if the pipe will be closed in the middle of writing, the HAL will receive EPIPE signal which will kill the process).

Patch-set: 1
Reviewer: Gerrit User 1042638 <1042638@85c56323-6fa9-3386-8a01-6480fb634889>