Update patch set 1
Patch Set 1:

Yeah, I think there are a couple other CLs that need to be moved to. Since configstore is a generic service used by many pieces of the Android operating system, I think it makes a lot of sense for it to be developed directly in AOSP in case bluetooth/wifi/telephony/etc.. need to rely on it.

Patch-set: 1