Update patch set 1
Patch Set 1:

It looks like what mm is currently doing is walking up the tree to find the first Android.mk that does exist and using that as its argument. I wonder if we want to update mm to pretend that an empty Android.mk exists in the current directory (if no Android.mk exists) or if users will be expecting its current behavior. I'm guessing we probably want to leave mm unchanged (although that would cause the meanings of the directories for `mm` and `mmm` to diverge, which could be odd, but I don't have any ideas about what we could do that would be less odd).

Patch-set: 1
Reviewer: Gerrit User 1137668 <1137668@85c56323-6fa9-3386-8a01-6480fb634889>