Add NFC HAL 1.1
Adds factoryReset(). This must be used by the HAL to clear the chip and
use the do a full initialization at the next init.

Adds closeForPowerOffCase(). There are vendor specific
configurations that are need during NFC power off to enable low battery
and power off use cases. This will distinguish cases where the user
turns NFC off because they do not want to use NFC and NFC is turned off
as a result of device switched off.

HCI_NETWORK_RESET event has been added to indicate something went wrong
on the firmware level and HCI network needs to be re-initialized.

Bug: 70294869
Bug: 70294551
Test: None
Change-Id: Ib981a56ac94e06b74bd901a159791f09ea16483a