Update patch set 6
Patch Set 6:

The data here needs to be structured.

Requirements for Structured Interfaces in Treble
All data passed through HIDL must be explicitly defined. We have adopted the following rules in order to make sure that an implementation and client will continue to work together even when compiled separately or developed on independently:
It can be described in HIDL directly (using structs/enums/etc..) with semantic names and meaning.
It can be described by a public standard (such as ISO/IEC 7816).
It can be described by a hardware standard or physical layout of hardware.
(only if necessary) It can be opaque data (for instance, public keys, ids, etc..). If these values are used, they must only ever be read by one side of the HIDL interface. For instance, if vendor.img code gives something on the system.img a “string message” or “vec<uint8_t> data” then that data may never be parsed by the system.img itself. It can only ever be passed back to the vendor code to interpret. This can always be fixed with a different data structure.
Rules of Thumb
For a well structured interface, adding new types of functionality that were not part of the original design should require a modification to the HIDL interface. See tl;dr above.
You should be able to write an implementation or client of a hal by looking at the .hal files alone even without looking at Android source (this includes looking at public standards).
All things should be for sure. Statements like “an implementation may do A or B” encourage implementations to become intertwined with the clients they are developed with. By spelling out required behavior exactly, this can be avoided.

Patch-set: 6
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>