Update patch set 8
Patch Set 8: Code-Review+1

This is the polled API for Location... it's also going to tie the bandwidth to specific cells which the other API does not (because it would need the information in CellInfo (the EARFCN and PCI to do that). Perhaps re-review the document and see what we agreed upon?

Patch-set: 8
Reviewer: Gerrit User 1060151 <1060151@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1