Update patch set 2
Patch Set 2: Code-Review-1

This is adding methods to an obsolete API. there is a new requestNetworkScan() API with significantly richer functionality that addresses most (if not all) of the needs of this along with a number of other enhancements and simplifications. I kindly request Sony to review the requestNetworkScan and see if there are any important gaps remaining between what is supported and what they would like.

Patch-set: 2
Reviewer: Gerrit User 1060151 <1060151@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1