Update patch set 16
Patch Set 16:

@Robert, I also think it should be consistent, so our two options are:
1) remove the field that was present in the old API and don't add its equivalent in the new API; or
2) add the field in the new API to mirror the functionality of the old API.

The only way to be consistent between DL and UL and also be consistent with the old API that we're extending is to do exactly what's proposed in the current CL. Neither Eric nor I dreamed up this "confidence" number. It's carrying forward the existing functionality in a logically consistent/equivalent manner.

Patch-set: 16
Reviewer: Gerrit User 1060151 <1060151@85c56323-6fa9-3386-8a01-6480fb634889>