Update patch set 2
Patch Set 2:

(7 comments)

Overall I think there's benefit in getting the device capabilities, but the proposed method seems both rigid in construction and narrow in scope. Can the design be made more flexible so that we don't need to add another one of these? Alternatively, can we name this API such that it fits within an obvious and narrow quarantine zone of "eUICC" lest anybody try to use the data in a generic fashion? Then we will still need another API to do something similar in the future.

Patch-set: 2
Reviewer: Gerrit User 1060151 <1060151@85c56323-6fa9-3386-8a01-6480fb634889>