Update patch set 2
Patch Set 2:

Next note: the encoding in section 4.2 indicates that all of this information is optional. Why do we need it? It seems like a lot of info to send to a server to provision a SIM.

Will something be broken if we don't provide it?

Patch-set: 2
Reviewer: Gerrit User 1060151 <1060151@85c56323-6fa9-3386-8a01-6480fb634889>