Update patch set 1
Patch Set 1:

AIUI the Nr standard at this time only specifies the use of 5G as a secondary serving cell, which means that it doesn't really do anything in Signal Strength. Signal strength is first-and-foremost for UI, which will display the signal strength of the camped/registered/primary serving cell, and if my understanding is correct the NR cell will never be the camped/registered/PSC.

I don't know how we'd even expect the UX to work in this case as we transition between IDLE and CONNECTED. Do we want the signal bars to use the 5g signal strength in connected and then go back to camped LTE cell in idle mode? That seems like something we'd want to avoid.

Where I think we would be adding information about secondary serving cells is in both the CellInfo classes and in the unsolicited Physical Channel Configuration Changed indication. Perhaps now would be a good time to circle back and consider support for that new API from the modem as the appropriate place to surface this information?

Patch-set: 1