Update patch set 2
Patch Set 2: Code-Review-1

> Patch Set 2: Presubmit-Ready+1
> 
> (2 comments)
> 
> > Patch Set 1:
> > 
> > (2 comments)
> > 
> > It would also be good to add a comment to requestNetworkScan() that says that only GERAN/UTRAN/EUTRAN are acceptable values for the ran type.
> 
> Per offline discussion, we actually lucked out since the 1.2 NetworkScanRequest uses the 1.1 RadioAccessNetworks. Updating the 1.2 RadioAccessNetwork doesn't contaminate the 1.1 type, so we're fine there.

I object to that.  The expectation is that the RadioAccessNetwork thing has been updated but is semantically the same thing and NetworkScanRequest should be or will be updated to take the 1.2 values.

If these aren't all equivalent we should make a separate enum.  Why cant CDMA2000 be returned by NetworkScanRequest?  That may explain how these are really different types and it should be fanned out.

Patch-set: 2
Reviewer: Gerrit User 1002609 <1002609@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1